Experimental: tune the transmitted TCP rx_win value
In order to stop large queues building we can control the rate that
packets are being sent to us by setting the TCP rx_win value
appropriately.

This patch calculates a value of rx_win based on the send window given
by the other side, and the other side's cwnd and mss.

This will almost certainly need further work, but at the moment is seems
to stop queues getting too long while ensuring that there are packets queued
ready to send when cwnd permits it.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>