Fix RTO calculation, and re-open a closed send window on other side
a) The value of RTO was being set too high resulting in delays recovering
from timeouts.

b) If the send queue on one side became full, the send window on the other
side was closed to stop being inundated by packets. However, when packets
were ack'd on the output side, the send window was not being reopened,
resulting in packet transfer stalling.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>