Reduce delayed ACK timer to take account of RTT
Quite a few retransmissions were being seen due to delayed ACKS. It
was waiting 200ms, which is the min timeout Linux uses for a tail
loss probe (RFC8985).

We need to allow for RTT when setting the delayed ack timer. Our best
estimate for the time a packet takes to arrive at the remote end is
RTT / 2. Allow two of these intervals to be safe, so subtract 1 * RTT
from the delayed ack timer.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>