When not queueing a packet set bit to indicate on private area
When the first SYN is forwarded it is not queued, therefore the
private area in the mbuf is not set up and after a successful burst
send the packet has been freed. We need to set the bit to indicate
that the packet has not been retained and the private area of the
mbuf no longer exists.

Currently this also applies to SYN+ACK and FIN packets, but they
need to be queued.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>