EXPERIMENTAL: Resend packets after rte_mbuf_tx_burst() failed to send them
Sometime immediately resending packets that rte_mbuf_tx_burst() failed
to send is successful. Even if not successful, we need to resend them
so we now have a queue of packets that failed to send, so that the next
time we send any packets we also resend the ones that failed.

This patch is EXPERIMENTAL since it might cause more problems than it
solves if we just keep sending more and more packets. It might be that we
need to gracefully back off resending packets, but they have to be sent
at some point.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>