Add option to send debug output to circular buffer only output on error
When enabled, this commit makes all printfs write to a 64Mb circular
buffer (in hugepages), and this is only written to stdout if a printf
format string starts or ends with the string ERROR.

The purpose of this is to avoid gigabytes of log output which has nothing
of interest, but when there is an error sufficient log entries from before
the error occurred will be written to stdout, thereby allowing a reasonable
chance of identifying the problem.

There is a runtime option option to write the output of the circular buffer
to stdout before the buffer overwrites itself. This has the benefit that
there is only one user/kernel/user space context switch per 64Mb of log
output, and opposed to a context switch every 4kb.

Whichever way this is used, there is a significant saving in system time
used.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>