Add option to check mbuf if DPDK built with RTE_LIBRTE_MEMPOOL_DEBUG
If an mbuf has been freed, its refcnt remains 1 on the basis that it
is referenced from the free queue (a refcnt of 0 is an error).

If DPDK is built with RTE_LIBRTE_MEMPOOL_DEBUG defined, it adds a
cookie field to the mbuf header, and we can check this to see if
an mbuf has been freed or not. If an mbuf is on an eflow pktlist
but has also been freed from DPDK's point of view, then there is a
bug.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>