Fix both freeing an mbuf and queueing it for sending
Due to an incorrect calculation of whether the SEQ of a packet
received after a packet with FIN set has been received was after
the SEQ of the FIN, the ACK of a FIN was both being forwarded
and freed.

This caused a problem with packets being freed twice, and DPDK
would then put two packets into the same mbuf, sometimes updating
the mbuf after TFO already had it queued. If the packet belonged
to a different TCP connection, mayhem ensued.

Signed-off-by: Quentin Armitage <quentin@armitage.org.uk>